#1: Move common video functions to video_utils.
